# Conflicts:
#	app/src/main/java/com/example/ibuy/BranchSelection.java
#	app/src/main/java/com/example/ibuy/LoginActivity.java
#	app/src/main/java/com/example/ibuy/MainActivity.java
#	app/src/main/res/layout/activity_branch_selection.xml
#	app/src/main/res/layout/activity_login.xml
#	app/src/main/res/layout/activity_register.xml

+++ app/src/main/java/com/example/ibuy/Counter.java
@@ -2,7 +2,7 @@ package com.example.ibuy;
 
 
 public class Counter {
-    private int counter;
+    private double counter;
 
     public Counter() {
         this.counter = counter;
@@ -25,7 +25,7 @@ public class Counter {
         this.counter = 0;
     }
 
-    public int getcounter() {
+    public double getcounter() {
         return this.counter;
     }
 }
